Adding lib/ObjectPath.js back in, correcting erroneous AMD module definition
